requirements.txt
streamlit==1.32.0
ultralytics==8.1.29
torch==2.2.2
torchvision==0.17.2
easyocr==1.7.1
opencv-python-headless==4.9.0.80
requests==2.31.0
beautifulsoup4==4.12.3
plotly==5.20.0
Pillow==10.2.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
scikit-image==0.22.0
reportlab==4.2.2
